Code cleanup that followed on from a bug-fix required for Sierra. I ended up removing the old CommCore, CommUtilsBase etc classes. This greatly simplifies the communication infrastructure.

+++ packages/fei/base/snl_fei_SubdMsgHandler.cpp
@@ -98,7 +98,7 @@ int snl_fei::SubdMsgHandler::processRecvMessage(int srcProc,
 
   int* msgPtr = &message[0];
 
-  if ((int)message.size() != ids->size()) {
+  if (message.size() != ids->size()) {
     ERReturn(-1);
   }
